+++ apps/react/src/challenges/inline-options/OptionType.tsx
@@ -1,0 +1,5 @@
+export type OptionType = {
+  id: string;
+  text: string;
+  isDisabled?: boolean;
+};
